[MeL] Move impl. of Properties into own files.
